+++ tests/testthat/test-bm_list.R
@@ -1,4 +1,7 @@
 test_that("bm_list()", {
+
+    expect_error(bm_list(2), "Some elements were not")
+
     plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
     plus_sign[5L, 3:7] <- 1L
     plus_sign[3:7, 5L] <- 1L